requirements.txt
Pillow>=10.0.0
pystray>=0.19.0
pyserial>=3.5
PyYAML>=6.0